refactor(protocol): CodecParams — Codec enum + bitrate: u32
Replace CodecParams.name: String with a Codec enum (single Opus
variant, serde "opus") so the type system enforces the only supported
codec. Change bitrate: i32 → u32 since negative bitrate is meaningless;
cast as i32 only at the audiopus boundary (OpusEncoder::new).

All constructors and tests updated. CLI and tauri command signatures
follow (bitrate arg becomes u32 throughout). No TS bindings change
(specta maps both to number).

+++ crates/splitter-cli/src/commands/loop_cmd.rs
@@ -9,7 +9,7 @@ use tokio::time::Duration;
 pub(crate) async fn run(
     input: &str,
     output: &str,
-    bitrate: i32,
+    bitrate: u32,
     source: crate::Source,
     fec_mode: crate::SendFecMode,
     simulated_loss_pct: u8,
@@ -28,7 +28,7 @@ pub(crate) async fn run(
     let core_fec_mode = map_fec_mode(fec_mode);
     let mut fec = splitter_core::net::fec::FecController::new(core_fec_mode, 1, 0, 10);
 
-    let mut enc = OpusEncoder::new(bitrate)?;
+    let mut enc = OpusEncoder::new(bitrate as i32)?;
     let mut dec = OpusDecoder::new()?;
     let mut payload = BytesMut::with_capacity(400);
     let mut frame = vec![0.0f32; FRAME_STEREO_SAMPLES];
